Move start_fpm to a specific file
