+++ src/main.rs
@@ -11,6 +11,7 @@ mod utils {
     pub(crate) mod current_process_name;
     pub(crate) mod php_binaries;
     pub(crate) mod php_server;
+    pub(crate) mod php_server_fpm;
 }
 
 use crate::commands::php_list::command_config as php_list_cmd;
